+++ game/core/src/com/eng1/game/Player.java
@@ -8,6 +8,7 @@ import com.badlogic.gdx.graphics.g2d.Batch;
 import com.badlogic.gdx.maps.tiled.TiledMapTileLayer;
 import com.badlogic.gdx.maps.tiled.TiledMapTileLayer.Cell;
 import com.badlogic.gdx.math.Vector2;
+import com.eng1.game.game.activity.Activity;
 import lombok.Getter;
 import lombok.Setter;
