+++ game/core/src/com/eng1/game/game/activity/Activity.java
@@ -1,5 +1,7 @@
-package com.eng1.game;
+package com.eng1.game.game.activity;
 
+import com.eng1.game.GameStats;
+import com.eng1.game.HeslingtonHustle;
 import com.eng1.game.screens.Screens;
 import lombok.Getter;
 import lombok.Setter;
