+++ game/core/src/com/eng1/game/screens/EndScreen.java
@@ -12,7 +12,7 @@ import com.badlogic.gdx.scenes.scene2d.ui.TextButton;
 import com.badlogic.gdx.scenes.scene2d.utils.ChangeListener;
 import com.badlogic.gdx.utils.viewport.ScreenViewport;
 
-import com.eng1.game.Activity;
+import com.eng1.game.game.activity.Activity;
 
 import java.util.Map;
